+++ internal/accrual/models/order.go
@@ -5,9 +5,9 @@ import (
 )
 
 type Order struct {
-	UUID    string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int64  `json:"accrual"`
+	UUID    string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
 
 type OrderItem struct {
